\name{getDIC}
\alias{getDIC}
\title{DIC and pD}
\usage{
  getDIC(jagsResults)
}
\arguments{
  \item{jagsResults}{An \code{\link[=rjags-class]{rjags}}
  result}
}
\value{
  A data frame with the DIC and pD
}
\description{
  Obtain the DIC and estimated number of parameters (pD)
  for an \code{\link[=rjags-class]{rjags}} fit.
}
\author{
  Chloe Bracis
}

