\name{Lamprey.Data}
\alias{Lamprey.Data}
\docType{data}
\title{Lamprey: Interpolated Data}
\description{
The original \code{\link[waddle]{Lamprey}} data was interpolated to a 2-min interval 
}
\usage{data(Lamprey.Data)}
\format{
  Data frame containing locations and times. See \code{\link[waddle]{Lamprey}}.
}
\source{
  \code{\link[waddle]{Lamprey}}}
\examples{
  data(Lamprey.Data)
}
\keyword{datasets}
