\name{WFR.Data}
\alias{WFR.Data}
\docType{data}
\title{Wild Forest Reindeer: Interpolated Data}
\description{
The original \code{\link[waddle]{WFR}} data was interpolated to a 2-day interval with the interpolated points jittered to avoid linear segments. 
}
\usage{data(WFR.Data)}
\format{
  Data frame containing locations and times. See \code{\link[waddle]{WFR}}.
}
\source{
\code{\link[waddle]{WFR}}}
\examples{
data(WFR.Data)
}
\keyword{datasets}
