\name{Wolf.Data}
\alias{Wolf.Data}
\docType{data}
\title{Wolf: Interpolated Data}
\description{
  The original \code{\link[waddle]{Wolf}} data was averaged to a 1-day interval. 
}
\usage{data(Wolf.Data)}
\format{
  Data frame containing locations and times. See \code{\link[waddle]{Wolf}}.
}
\source{
  \code{\link[waddle]{Wolf}}}
\examples{
  data(Wolf.Data)
}
\keyword{datasets}
